Remove redundent inheritance of DAU, SA Solvers
In the last version, the DAUSolver class and SASolver class inherit
the Solver and their core algorithm respectively. The contents of DAUSolver
and SASolver are nearly the same. In this version, I extend the Solver
class to remove redundent code and also to make the API clear.

When user want to write a new algorithm as a solver. They just need to
inherit the Solver class and override the solve method, use the
self.logger to log out the algorithm message. The solve method in the
Solver class is literally a vertual function and it returns two
dataframes, one is the shift data and another is the algorithm data.
Both of returned data will be display on the GUI. The Solver class has
three Signal variables, two of which are used in this version. One is
the finished and the other is the error. The finished connect to the
WorkingArea and will emit the results of algorithm to display on the
GUI. While the error will display the exception message shown by
QMessageBox.